fix error

+++ FDF_NEW/fdf.h
@@ -3,6 +3,7 @@
 
 #include <unistd.h>
 #include <stdlib.h>
+#include <stdio.h>
 #include "libft/libft.h"
 #include "get_next_line/get_next_line.h"
 #include "ft_printf/ft_printf.h"
@@ -19,6 +20,6 @@ typedef struct
 }				fdf;
 
 void	read_file(char *file_name, fdf *data);
-void	bresenham(int x, int y, int x1, int y1, fdf *data);
+void	bresenham(float x, float y, float x1, float y1, fdf *data);
 
 #endif
